requirements.txt
torch
torchvision
tensorboard
tensorboardX
pillow
